Skip test on GUN_UNUSUAL_DESIGN since they can't be inspected

+++ code/modules/asset_cache/asset_list_items.dm
@@ -380,6 +380,8 @@
 	for(var/obj/item/weapon/gun/current_gun as anything in subtypesof(/obj/item/weapon/gun))
 		if(isnull(initial(current_gun.icon_state)))
 			continue
+		if(initial(current_gun.flags_gun_features) & GUN_UNUSUAL_DESIGN)
+			continue // These don't have a way to inspect weapon stats
 		var/obj/item/weapon/gun/temp_gun = new current_gun
 		var/icon_state = temp_gun.base_gun_icon // base_gun_icon is set in Initialize generally
 		qdel(temp_gun)
